customer-managment/app/requirements.txt
flask-restful
flask
Flask==1.1.2
werkzeug==1.0.1
Flask-Migrate
Flask-Script
Flask-RestPlus
pymongo
kafka-python
json-logging